Move sample code to sample.go
We need to add more logic to optimize the sample for ODF. Keeping the
code related to create a sample in separate file makes this easier to
follow, and makes the code to create a config file cleaner.
