+++ pkg/config/config.go
@@ -4,12 +4,9 @@
 package config
 
 import (
-	"bytes"
-	_ "embed"
 	"errors"
 	"fmt"
 	"os"
-	"text/template"
 
 	"github.com/ramendr/ramen/e2e/config"
 	"github.com/ramendr/ramen/e2e/deployers"
@@ -17,9 +14,6 @@ import (
 	"github.com/ramendr/ramen/e2e/workloads"
 	"github.com/ramendr/ramenctl/pkg/console"
 )
-
-//go:embed sample.yaml
-var sampleConfig string
 
 func CreateSampleConfig(filename, commandName, envFile string) error {
 	var sample *Sample
@@ -60,25 +54,6 @@ func ReadConfig(filename string) (*types.Config, error) {
 	return config, nil
 }
 
-type Sample struct {
-	CommandName         string
-	HubKubeconfig       string
-	PrimaryKubeconfig   string
-	SecondaryKubeconfig string
-}
-
-func (s *Sample) Bytes() ([]byte, error) {
-	t, err := template.New("sample").Parse(sampleConfig)
-	if err != nil {
-		return nil, err
-	}
-	var buf bytes.Buffer
-	if err := t.Execute(&buf, s); err != nil {
-		return nil, err
-	}
-	return buf.Bytes(), nil
-}
-
 func createFile(name string, content []byte) error {
 	f, err := os.OpenFile(name, os.O_CREATE|os.O_EXCL|os.O_RDWR, 0o600)
 	if err != nil {
@@ -89,26 +64,4 @@ func createFile(name string, content []byte) error {
 		return err
 	}
 	return f.Close()
-}
-
-func defaultSample(commandName string) *Sample {
-	return &Sample{
-		CommandName:         commandName,
-		HubKubeconfig:       "hub/config",
-		PrimaryKubeconfig:   "primary/config",
-		SecondaryKubeconfig: "secondary/config",
-	}
-}
-
-func sampleFromEnvFile(envFile, commandName string) (*Sample, error) {
-	env, err := ReadEnvFile(envFile)
-	if err != nil {
-		return nil, err
-	}
-	return &Sample{
-		CommandName:         commandName,
-		HubKubeconfig:       env.KubeconfigPath(env.Ramen.Hub),
-		PrimaryKubeconfig:   env.KubeconfigPath(env.Ramen.Clusters[0]),
-		SecondaryKubeconfig: env.KubeconfigPath(env.Ramen.Clusters[1]),
-	}, nil
 }
